+++ source/games/duke/src/game_misc.cpp
@@ -234,7 +234,7 @@ void V_AddBlend (float r, float g, float b, float a, float v_blend[4])
 //
 //---------------------------------------------------------------------------
 
-void displayrest(int smoothratio)
+void displayrest(double smoothratio)
 {
 	int i, j;
 	unsigned char fader = 0, fadeg = 0, fadeb = 0, fadef = 0, tintr = 0, tintg = 0, tintb = 0, tintf = 0, dotint = 0;
